+++ enterprise/admin/forms.py
@@ -395,6 +395,7 @@ class EnterpriseCustomerAdminForm(forms.ModelForm):
             "enable_audit_data_reporting",
             "replace_sensitive_sso_username",
             "hide_course_original_price",
+            "disable_expiry_messaging_for_learner_credit",
             "enable_portal_code_management_screen",
             "enable_portal_subscription_management_screen",
             "enable_learner_portal",
@@ -767,10 +768,9 @@ class AdminNotificationForm(forms.ModelForm):
         if expiration_date < start_date:
             # `start_date` must always come before the `expiration_date`
             raise ValidationError({'expiration_date': ['Expiration date should be after start date.']})
-        # pylint: disable=unsupported-binary-operation
         admin_notification = AdminNotification.objects.filter(
-            Q(start_date__range=(start_date, expiration_date)) |
-            Q(expiration_date__range=(start_date, expiration_date)) |
+            Q(start_date__range=(start_date, expiration_date)) |  # pylint: disable=unsupported-binary-operation
+            Q(expiration_date__range=(start_date, expiration_date)) |  # pylint: disable=unsupported-binary-operation
             Q(start_date__lt=start_date, expiration_date__gt=expiration_date)
         ).exclude(
             pk=self.instance.id