+++ enterprise/signals.py
@@ -11,12 +11,13 @@ from enterprise import models, roles_api
 from enterprise.api import activate_admin_permissions
 from enterprise.api_client.enterprise_catalog import EnterpriseCatalogApiClient
 from enterprise.decorators import disable_for_loaddata
-from enterprise.tasks import create_enterprise_enrollment, update_enterprise_learners_user_preference
+from enterprise.tasks import create_enterprise_enrollment
 from enterprise.utils import (
     NotConnectedToOpenEdX,
     get_default_catalog_content_filter,
     localized_utcnow,
     unset_enterprise_learner_language,
+    unset_language_of_all_enterprise_learners,
 )
 from integrated_channels.blackboard.models import BlackboardEnterpriseCustomerConfiguration
 from integrated_channels.canvas.models import CanvasEnterpriseCustomerConfiguration
@@ -103,11 +104,10 @@ def update_lang_pref_of_all_learners(sender, instance, **kwargs):  # pylint: dis
     # The middleware in the enterprise will handle the cases for setting a proper language for the learner.
     if instance.default_language:
         prev_state = models.EnterpriseCustomer.objects.filter(uuid=instance.uuid).first()
-        if prev_state and prev_state.default_language != instance.default_language:
+        if prev_state is None or prev_state.default_language != instance.default_language:
             # Unset the language preference of all the learners linked with the enterprise customer.
             # The middleware in the enterprise will handle the cases for setting a proper language for the learner.
-            logger.info('Task triggered to update user preference for learners. Enterprise: [%s]', instance.uuid)
-            update_enterprise_learners_user_preference.delay(instance.uuid)
+            unset_language_of_all_enterprise_learners(instance)
 
 
 @receiver(pre_save, sender=models.EnterpriseCustomerBrandingConfiguration)