+++ enterprise/utils.py
@@ -7,6 +7,7 @@ import json
 import os
 import re
 from collections import OrderedDict
+from itertools import islice
 from urllib.parse import parse_qs, quote, urlencode, urljoin, urlparse, urlsplit, urlunsplit
 from uuid import UUID, uuid4
 
@@ -2219,18 +2220,17 @@ def get_platform_logo_url():
     return urljoin(settings.LMS_ROOT_URL, logo_url)
 
 
-def unset_language_of_all_enterprise_learners(enterprise_customer_uuid):
+def unset_language_of_all_enterprise_learners(enterprise_customer):
     """
     Unset the language preference of all the learners belonging to the given enterprise customer.
 
     Arguments:
-        enterprise_customer_uuid (UUI): uuid of an enterprise customer
+        enterprise_customer (UUI): Instance of the enterprise customer.
     """
     if UserPreference:
-        enterprise_customer = get_enterprise_customer(enterprise_customer_uuid)
         user_ids = list(enterprise_customer.enterprise_customer_users.values_list('user_id', flat=True))
 
-        LOGGER.info('Update user preference started for learners. Enterprise: [%s]', enterprise_customer_uuid)
+        LOGGER.info('Update user preference started for learners. Enterprise: [%s]', enterprise_customer.uuid)
 
         for chunk in batch(user_ids, batch_size=10000):
             UserPreference.objects.filter(
@@ -2241,7 +2241,7 @@ def unset_language_of_all_enterprise_learners(enterprise_customer_uuid):
             )
             LOGGER.info('Updated user preference for learners. Batch Size: [%s]', len(chunk))
 
-        LOGGER.info('Update user preference completed for learners. Enterprise: [%s]', enterprise_customer_uuid)
+        LOGGER.info('Update user preference completed for learners. Enterprise: [%s]', enterprise_customer.uuid)
 
 
 def unset_enterprise_learner_language(enterprise_customer_user):
@@ -2296,6 +2296,22 @@ def batch(iterable, batch_size=1):
         iterable_len = len(iterable)
     for index in range(0, iterable_len, batch_size):
         yield iterable[index:min(index + batch_size, iterable_len)]
+
+
+def batch_dict(dict_data, chunk_size=1):
+    """
+    Breaks up a dictionary into equal-sized chunks.
+    No fillers values are added for any 'remainder' chunks
+
+    Arguments:
+        dict (dict): A dictionary to chunk
+        chunk_size (int): the size of each chunk. Defaults to 1.
+    Returns:
+        generator: iterates through each chunk of a dictionary
+    """
+    it = iter(dict_data.items())
+    for _ in range(0, len(dict_data), chunk_size):
+        yield dict(islice(it, chunk_size))
 
 
 def get_best_mode_from_course_key(course_key):