+++ tests/test_integrated_channels/test_blackboard/test_client.py
@@ -37,8 +37,8 @@ class TestBlackboardApiClient(unittest.TestCase):
         self.token = 'token'
         self.global_config = BlackboardGlobalConfigurationFactory()
         self.enterprise_config = BlackboardEnterpriseCustomerConfigurationFactory(
-            client_id='id',
-            client_secret='secret',
+            decrypted_client_id='id',
+            decrypted_client_secret='secret',
             blackboard_base_url='https://base.url',
             refresh_token=self.token,
         )
@@ -67,8 +67,8 @@ class TestBlackboardApiClient(unittest.TestCase):
 
     def test_client_pulls_auth_creds_from_global_if_not_found(self):
         enterprise_config = BlackboardEnterpriseCustomerConfigurationFactory(
-            client_id='',
-            client_secret='',
+            decrypted_client_id='',
+            decrypted_client_secret='',
         )
         client = BlackboardAPIClient(enterprise_config)
         auth_header = client._create_auth_header()
@@ -78,8 +78,8 @@ class TestBlackboardApiClient(unittest.TestCase):
 
     def test_oauth_absent_refresh_token_fails(self):
         enterprise_config = BlackboardEnterpriseCustomerConfigurationFactory(
-            client_id='id2',
-            client_secret='secret',
+            decrypted_client_id='id2',
+            decrypted_client_secret='secret',
             blackboard_base_url='https://base.url.2',
             refresh_token='',
         )
@@ -94,8 +94,8 @@ class TestBlackboardApiClient(unittest.TestCase):
         a valid refresh_token is used to replace it, and access_token is obtained
         """
         enterprise_config = BlackboardEnterpriseCustomerConfigurationFactory(
-            client_id='id3',
-            client_secret='secret',
+            decrypted_client_id='id3',
+            decrypted_client_secret='secret',
             blackboard_base_url='https://base.url.3',
             refresh_token='a-token',
         )